LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/visual-scrollIntoView-002-expected.txt



FAIL Element.scrollIntoView scrolls visually to a position: fixed element with non-zero layout scroll offset assert_greater_than: expected a number greater than 1 but got 1
